refactor(nft mint): reorganize components, fix NFT display
